requirements/base.txt
# Django
Django==1.11.6

# Environmental var configuration
django-configurations==2.0

# For the persistence stores
psycopg2==2.7.3.1
dj-database-url==0.4.2

# Models
django-model-utils==3.0.0

# Rest apis
djangorestframework==3.7.1
Markdown==2.6.9
django-filter==1.1.0
djangorestframework-jwt==1.11.0

# Static files
whitenoise==3.3.1

# Media files
boto==2.48.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.8.1

# For asynchronous queuing
django-rq==0.9.6

# Time zones support
pytz==2017.2
